+++ tests/Ireland/NewYearsDayTest.php
@@ -50,7 +50,7 @@ class NewYearsDayTest extends IrelandBaseTestCase implements HolidayTestCase
         $date = new \DateTime($expected, new \DateTimeZone(self::TIMEZONE));
         $this->assertHoliday(self::REGION, self::HOLIDAY, $year, $date);
 
-        if (0 === (int) $date->format('w')) {
+        if (\in_array((int) $date->format('w'), [0, 6], true)) {
             $date->modify('next monday');
             $this->assertHoliday(self::REGION, 'substituteHoliday:' . self::HOLIDAY, $year, $date);
         }